backend/requirements.txt
opencv-python
numpy
librosa
soundfile
ffmpeg-python
pydantic
tqdm